Uploaded Version 1
https://www.unb.ca/cic/datasets/nsl.html
You can find the dataset from the above URL mentioned.

Here, NSL-KDD is the main file in which we first did our preprocessing (Exploratory Data Analysis) on the dataset to make dataset applicable for the classifiers to run. Then we applied different Machine LEarning Algorithms to achieve the better accuracy and F1 score to find out the anomalies in the dataset. 
ANN file is the code file for the Artificial Neural Based model, we have also done the hyperparameter tuning with the help of Optuna.
All Results are basically loading all the classifiers joblib files which were obtained after training the classifiers.